6 Absorption models built.
